requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=py38h06a4308_0
addict=2.4.0=pypi_0
aiohttp=3.9.5=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.3=py38h06a4308_0
attrs=23.1.0=py38h06a4308_0
basicsr=1.3.4.9=pypi_0
blas=1.0=mkl
blinker=1.6.2=py38h06a4308_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.3.3=py38h06a4308_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=py38h5eee18b_1
charset-normalizer=2.0.4=pypi_0
click=8.1.7=py38h06a4308_0
cmake=3.30.2=pypi_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=42.0.5=py38hdda0065_1
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
et-xmlfile=2.0.0=pypi_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py38h5eee18b_0
fsspec=2024.12.0=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.29.0=py38h06a4308_0
google-auth-oauthlib=0.5.2=py38h06a4308_0
grpcio=1.62.2=py38h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
huggingface-hub=0.27.1=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.0.1=py38h06a4308_0
importlib_resources=6.4.0=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
jinja2=3.1.3=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.11.0.15=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.37=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
lit=18.1.8=pypi_0
lmdb=1.5.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.4.1=py38h06a4308_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.0.4=py38h5eee18b_0
mysql=5.7.24=h721c034_2
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=py38h06a4308_0
openai-clip=1.0.1=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openpyxl=3.1.5=pypi_0
openssl=3.0.14=h5eee18b_0
packaging=24.0=pypi_0
pcre2=10.42=hebb0a14_1
pillow=10.2.0=pypi_0
pip=23.3.1=pypi_0
platformdirs=4.3.6=pypi_0
ply=3.11=py38_0
portalocker=2.10.1=pypi_0
protobuf=4.25.3=py38h12ddb61_0
psutil=6.1.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.8.0=py38h06a4308_0
pyopenssl=24.2.1=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0post0=py38h06a4308_2
pytorch=2.0.0=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
qt-main=5.15.2=h53bd1ea_10
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2024.4.16=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=py38h06a4308_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.5.2=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
segment-anything=1.0=dev_0
sentry-sdk=2.18.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=68.2.2=pypi_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py38h06a4308_0
tabulate=0.9.0=pypi_0
tb-nightly=2.14.0a20230808=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=py38h06a4308_0
tensorboard-data-server=0.7.0=py38h52d8a92_1
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.4.0=pypi_0
terminaltables=3.1.10=pypi_0
tifffile=2023.7.10=pypi_0
timm=1.0.12=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py38h06a4308_0
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu118
tornado=6.4.1=py38h5eee18b_0
tqdm=4.66.2=pypi_0
typing-extensions=4.9.0=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.1.0=pypi_0
wandb=0.18.7=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=py38h06a4308_0
wheel=0.41.2=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.43.0=pypi_0
yarl=1.9.3=py38h5eee18b_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
